separate and clean code
