--- config.py
@@ -1,8 +1,0 @@
-
-
-
-config={
-#'ChatOpenAI' 'huggingCustomEndpoint' 'wizardVicuna13_local' 'vicuna13b_replicate'  'falcon7b_octoAI' 'flan-t5-xxl'
-"LLM_Name":'vicuna13b_replicate'
-
-}
