+++ src/config.py
@@ -1,0 +1,12 @@
+
+LLMList=['ChatOpenAI',
+         'huggingCustomEndpoint',
+         'wizardVicuna13_local',
+         'vicuna13b_replicate',
+         'falcon7b_octoAI',
+         'flan-t5-xxl',
+         'llama_v2_13b_replicate']
+
+config={
+"LLM_Name": LLMList[6]
+}
